Fixed bug 1021 (misplaced HAVE_SIGNAL_H condition in playmus.c and playwave.c)
 Ozkan Sezer      2010-10-08 01:45:34 PDT

The HAVE_SIGNAL_H dependent signal.h includes in playmus.c and playwave.c are
misplaced, because SDL_config.h is not included at that point and this causes
signal.h to be never included. Moving it after the SDL.h and SDL_mixer.h
includes fixes the issue, like in the following.